refactor: streamlined package structure for commands and renderers
GH-96
