+++ src/ossiq/ui/renderers/html/__init__.py
@@ -1,0 +1,1 @@
+"""Html command renderers"""
