+++ src/ossiq/ui/renderers/info/__init__.py
@@ -1,0 +1,1 @@
+"""Info command renderers"""
